Fix hamburger menu
